feat: wrapped function
